+++ src/store/mod.rs
@@ -23,8 +23,8 @@ where
     &mut self,
     alias: String,
     executor: String,
-    start_time: u128,
-    end_time: Option<u128>,
+    start_time: i64,
+    end_time: Option<i64>,
     job: Box<dyn Work>,
   ) -> Result<(), String>;
